http: move response helper to util
